+++ components/meals/index.ts
@@ -1,4 +1,3 @@
-export { default as MealPlanner } from './MealPlanner';
 export { default as PantrySidebar } from './PantrySidebar';
 export { default as ShoppingListSidebar } from './ShoppingListSidebar';
 export { default as RecipesSidebar } from './RecipesSidebar';