+++ utils/index.ts
@@ -20,10 +20,3 @@ export * from './taskUtils';
 
 // Ingredient utilities
 export * from './ingredientUtils';
-
-// Habits utilities
-export { HabitsStorage } from './habitsStorage';
-
-// Activities utilities
-export { ActivitiesStorage, ActivitiesListStorage } from './activitiesStorage';
-export type { ActivitiesByDate, ActivityItem } from './activitiesStorage';